pyproject.toml
[project]
name = "strands-pg"
version = "0.1.0"
description = "Postgres-backed session, memory, and API primitives for Strands agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Alice Rider", email = "brian.peterson@cloudshift.cc" },
]
dependencies = [
    "strands-agents>=1.13.0",
    "psycopg[binary,pool]>=3.2",
    "pgvector>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
]

[project.optional-dependencies]
bedrock = ["boto3>=1.35"]
ollama = ["ollama>=0.3"]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "httpx>=0.27",
]

[project.scripts]
strands-pg-migrate = "strands_pg.migrate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strands_pg"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
